generate api reports
