+++ plugins/search-backend-module-pg/alpha-api-report.md
@@ -1,0 +1,12 @@
+## API Report File for "@backstage/plugin-search-backend-module-pg"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { BackendFeature } from '@backstage/backend-plugin-api';
+
+// @alpha
+export const pgSearchEngineModule: () => BackendFeature;
+
+// (No @packageDocumentation comment for this package)
+```
