+++ plugins/search-backend-node/alpha-api-report.md
@@ -1,0 +1,62 @@
+## API Report File for "@backstage/plugin-search-backend-node"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { DocumentCollatorFactory } from '@backstage/plugin-search-common';
+import { DocumentDecoratorFactory } from '@backstage/plugin-search-common';
+import { DocumentTypeInfo } from '@backstage/plugin-search-common';
+import { ExtensionPoint } from '@backstage/backend-plugin-api';
+import { SearchEngine } from '@backstage/plugin-search-common';
+import { ServiceRef } from '@backstage/backend-plugin-api';
+import { TaskRunner } from '@backstage/backend-tasks';
+
+// @alpha
+export type IndexServiceBuildOptions = {
+  searchEngine: SearchEngine;
+  collators: RegisterCollatorParameters[];
+  decorators: RegisterDecoratorParameters[];
+};
+
+// @public
+export interface RegisterCollatorParameters {
+  factory: DocumentCollatorFactory;
+  schedule: TaskRunner;
+}
+
+// @public
+export interface RegisterDecoratorParameters {
+  factory: DocumentDecoratorFactory;
+}
+
+// @alpha
+export interface SearchEngineRegistryExtensionPoint {
+  // (undocumented)
+  setSearchEngine(searchEngine: SearchEngine): void;
+}
+
+// @alpha
+export const searchEngineRegistryExtensionPoint: ExtensionPoint<SearchEngineRegistryExtensionPoint>;
+
+// @alpha
+export interface SearchIndexRegistryExtensionPoint {
+  // (undocumented)
+  addCollator(options: RegisterCollatorParameters): void;
+  // (undocumented)
+  addDecorator(options: RegisterDecoratorParameters): void;
+}
+
+// @alpha
+export const searchIndexRegistryExtensionPoint: ExtensionPoint<SearchIndexRegistryExtensionPoint>;
+
+// @alpha
+export interface SearchIndexService {
+  getDocumentTypes(): Record<string, DocumentTypeInfo>;
+  start(options: IndexServiceBuildOptions): Promise<void>;
+}
+
+// @alpha
+export const searchIndexService: ServiceRef<SearchIndexService, 'plugin'>;
+
+// (No @packageDocumentation comment for this package)
+```
