+++ plugins/search-backend-node/src/index.ts
@@ -23,7 +23,7 @@
 export { IndexBuilder } from './IndexBuilder';
 export { Scheduler } from './Scheduler';
 export * from './collators';
-export { LunrSearchEngine, lunrSearchEngineModule } from './engines';
+export { LunrSearchEngine } from './engines';
 export type {
   ConcreteLunrQuery,
   LunrQueryTranslator,
