+++ plugins/search-backend/alpha-api-report.md
@@ -1,0 +1,12 @@
+## API Report File for "@backstage/plugin-search-backend"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { BackendFeature } from '@backstage/backend-plugin-api';
+
+// @alpha
+export const searchPlugin: () => BackendFeature;
+
+// (No @packageDocumentation comment for this package)
+```
